Break test_PDB.py in multiple smaller, modular, tests. (#3220)
* Simplified numpy imports

* Isolated Bio.PDB DSSP tests

* Isolated *Exposure tests

* Isolated ResidueDepth and MSMS tests

* Replaced function + global var by setUpClass

* Removed writer tests (PDBIO/MMCIFIO)

* Removed more writer tests (PDBIO)

* Isolated tests for PDBIO

* Isolated tests for MMCIFIO

* Removed unused function

* Missed one test before

* Isolated tests for PPBuilder and CaPPBuilder

* More cleanup...

* Use raise instead of self.skipTest

* Isolated PDBParser tests

* Isolated SMCRA tests

* Moved header tests to the appropriate file

* Silenced warnings in a couple of tests with QUIET=1

* Commented out time-consuming test

* Trigger
