+++ Tests/test_PDB_StructureAlignment.py
@@ -21,15 +21,14 @@ class StructureAlignTests(unittest.TestCase):
 
     def test_StructAlign(self):
         """Tests on module to align two proteins according to a FASTA alignment file."""
+        p = PDBParser(QUIET=1)
+
         al_file = "PDB/alignment_file.fa"
-        pdb2 = "PDB/1A8O.pdb"
-        pdb1 = "PDB/2XHE.pdb"
         with open(al_file) as handle:
             records = AlignIO.read(handle, "fasta")
-        p = PDBParser()
-        s1 = p.get_structure("1", pdb1)
-        p = PDBParser()
-        s2 = p.get_structure("2", pdb2)
+
+        s1 = p.get_structure("1", "PDB/2XHE.pdb")
+        s2 = p.get_structure("2", "PDB/1A8O.pdb")
         m1 = s1[0]
         m2 = s2[0]
         al = StructureAlignment(records, m1, m2)
